requirements.txt
python-consul
shortuuid
netifaces
loguru
